break out functionality into a separate module
