+++ Gemfile
@@ -7,3 +7,4 @@ git_source(:github) {|repo_name| "https://github.com/#{repo_name}" }
 gem "rspec"
 gem "rspec-core"
 gem "rake"
+gem "pry"
